src/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-openai
chromadb
python-dotenv
openai
tiktoken
